refactor: Relocate UBA type predicates + test
