+++ src/typeguards/index.ts
@@ -1,1 +1,2 @@
 export * from "./error";
+export * from "./UBA";
